+++ src/database/prisma.service.ts
@@ -1,0 +1,20 @@
+import { Injectable } from '@nestjs/common';
+import { ConfigService } from '@nestjs/config';
+import { PrismaClient } from '@prisma/client';
+
+@Injectable()
+export class PrismaService extends PrismaClient {
+  constructor(configService: ConfigService) {
+    super({
+      datasources: {
+        db: {
+          url: configService.getOrThrow<string>('DATABASE_URL'),
+        },
+      },
+      log:
+        configService.get<string>('NODE_ENV') === 'development'
+          ? ['warn', 'error']
+          : ['error'],
+    });
+  }
+}